change symbols and keywords to store whether to be quoted
That would be faster for printing? Although it will require more
memory, so not really sure. Anyway the idea was to use the same logic
used in the parsing. OH: actually wrong, it could store the need to be
quoted just because the first user quoted it!

 > '|a|
 |a|
 > 'a
 |a|

So, what really should do is still separate out the logic, then upon
first construction of the symbol calculate it. That way the
calculation would still be re-used (both 'lexically' and
'dynamically').