lily: solve the typeof issue via manual type declaration
Works, but is this a confusing error message?

 > (/ 1/9223372036854775807 2)
 ERR: #<int64-underflow-error (* 9223372036854775807 2)>

What else would I want to show, though? Would complicate the contents
of the exception object (not just 1 operation anymore). And perhaps
the above is just fine: you can see that it's a multiplication, but it
states underflow, so it must be a multiplication in the context of a
denominator.